[REFACTOR][#77] settings 내의 view 각각 분리
